Refactor worker helper functions
Separating the affiliation of functions between pulpcore.tasking.task and
pulpcore.tasking.utils (now _utils) to reflect which of those are used to
interact with tasks from the app level and which are internal plumbing of the
task worker.

[noissue]
