+++ pulpcore/tasking/tasks.py
@@ -1,23 +1,20 @@
-import logging
-from gettext import gettext as _
 import asyncio
 import contextlib
 import contextvars
 import importlib
-import traceback
+import logging
 import sys
+import traceback
+from gettext import gettext as _
 
-from django.db import transaction, connection
-from django.db.models import Model, Q
-from django.utils import timezone
-from django_guid import get_guid, set_guid
-from django_guid.utils import generate_guid
-
+from django.db import connection, transaction
+from django.db.models import Model
+from django_guid import get_guid
 from pulpcore.app.apps import MODULE_PLUGIN_VERSIONS
 from pulpcore.app.loggers import deprecation_logger
-from pulpcore.app.models import Task, TaskSchedule
-from pulpcore.app.util import get_url, get_domain, current_task
-from pulpcore.constants import TASK_FINAL_STATES, TASK_STATES, TASK_INCOMPLETE_STATES
+from pulpcore.app.models import Task
+from pulpcore.app.util import current_task, get_domain, get_url
+from pulpcore.constants import TASK_FINAL_STATES, TASK_INCOMPLETE_STATES, TASK_STATES
 
 _logger = logging.getLogger(__name__)
 
@@ -37,7 +34,7 @@ def _validate_and_get_resources(resources):
     return list(resource_set)
 
 
-def _wakeup_worker():
+def wakeup_worker():
     # Notify workers
     with connection.connection.cursor() as cursor:
         cursor.execute("NOTIFY pulp_worker_wakeup")
@@ -204,41 +201,40 @@ def dispatch(
                 task.set_canceling()
                 task.set_canceled(TASK_STATES.CANCELED, "Resources temporarily unavailable.")
     if notify_workers:
-        _wakeup_worker()
+        wakeup_worker()
     return task
 
 
-def dispatch_scheduled_tasks():
-    # Warning, dispatch_scheduled_tasks is not race condition free!
-    now = timezone.now()
-    # Dispatch all tasks old enough and not still running
-    for task_schedule in TaskSchedule.objects.filter(next_dispatch__lte=now).filter(
-        Q(last_task=None) | Q(last_task__state__in=TASK_FINAL_STATES)
-    ):
-        try:
-            if task_schedule.dispatch_interval is None:
-                # This was a timed one shot task schedule
-                task_schedule.next_dispatch = None
-            else:
-                # This is a recurring task schedule
-                while task_schedule.next_dispatch < now:
-                    # Do not schedule in the past
-                    task_schedule.next_dispatch += task_schedule.dispatch_interval
-            set_guid(generate_guid())
-            with transaction.atomic():
-                task_schedule.last_task = dispatch(
-                    task_schedule.task_name,
-                )
-                task_schedule.save(update_fields=["next_dispatch", "last_task"])
+def cancel_task(task_id):
+    """
+    Cancel the task that is represented by the given task_id.
 
-            _logger.info(
-                "Dispatched scheduled task {task_name} as task id {task_id}".format(
-                    task_name=task_schedule.task_name, task_id=task_schedule.last_task.pk
-                )
+    This method cancels only the task with given task_id, not the spawned tasks. This also updates
+    task's state to 'canceling'.
+
+    Args:
+        task_id (str): The ID of the task you wish to cancel
+
+    Raises:
+        rest_framework.exceptions.NotFound: If a task with given task_id does not exist
+    """
+    task = Task.objects.get(pk=task_id)
+
+    if task.state in TASK_FINAL_STATES:
+        # If the task is already done, just stop
+        _logger.debug(
+            "Task [{task_id}] already in a final state: {state}".format(
+                task_id=task_id, state=task.state
             )
-        except Exception as e:
-            _logger.warning(
-                "Dispatching scheduled task {task_name} failed. {error}".format(
-                    task_name=task_schedule.task_name, error=str(e)
-                )
-            )
+        )
+        return task
+
+    _logger.info(_("Canceling task: {id}").format(id=task_id))
+
+    # This is the only valid transition without holding the task lock
+    task.set_canceling()
+    # Notify the worker that might be running that task and other workers to clean up
+    with connection.cursor() as cursor:
+        cursor.execute("SELECT pg_notify('pulp_worker_cancel', %s)", (str(task.pk),))
+        cursor.execute("NOTIFY pulp_worker_wakeup")
+    return task
